+++ compiler-core/src/docs/tests.rs
@@ -96,6 +96,7 @@ fn compile_with_markdown_pages(
             title: (*title).into(),
             path: format!("{title}.html"),
             source: format!("{title}.md").into(),
+            unknown: Default::default(),
         })
         .collect_vec();
 
@@ -605,6 +606,7 @@ fn source_link_for_github_repository() {
         repo: "wobble".to_string(),
         path: None,
         tag_prefix: None,
+        unknown: Default::default(),
     });
 
     let modules = vec![("app.gleam", "pub type Wibble = Int")];
@@ -623,6 +625,7 @@ fn source_link_for_github_repository_with_path_and_tag_prefix() {
         repo: "wobble".to_string(),
         path: Some("path/to/package".to_string()),
         tag_prefix: Some("subdir-".into()),
+        unknown: Default::default(),
     });
 
     let modules = vec![("app.gleam", "pub type Wibble = Int")];
@@ -1219,6 +1222,7 @@ fn gitea_repository_url_has_no_double_slash() {
         repo: "forgejo_bug".into(),
         path: None,
         tag_prefix: None,
+        unknown: Default::default(),
     };
 
     assert_eq!(repo.url(), "https://code.example.org/person/forgejo_bug");
@@ -1245,6 +1249,7 @@ fn forgejo_single_line_definition() {
         repo: "wobble".into(),
         path: None,
         tag_prefix: None,
+        unknown: Default::default(),
     };
 
     config.name = EcoString::from("test_project_name");
@@ -1267,6 +1272,7 @@ fn forgejo_multiple_line_definition() {
         repo: "wobble".into(),
         path: None,
         tag_prefix: None,
+        unknown: Default::default(),
     };
 
     config.name = EcoString::from("test_project_name");